Prevent truncation ever being less than 1
When there are more x-axis labels than there are pixels of space the division will result in `truncation` being equal to zero. This causes truncation of x-axis labels to be completely bypassed as the `truncate()` function ignores truncations that are equal to 0. 

This patch makes 1 the minimum which means that in the worst case all the x-axis labels will be displayed as '...'. I think this is preferable to having the full labels positioned on top of one another.