ADDS HOOK FOR XML MODIFICATION PRE-RENDERING
An XML filter can be defined as a function taking a single argument,
supplied as the 'svg' atribute of the graph (an lxml etree?) and
returning a modified version/copy thereof.

XML filters can be added to a plot using the new plot method
add_xml_filter(func).

The hook that executes any filters added to the plot, executes them in
order immediately prior to conversion to a string, in the class method
svg.render, and also immediately prior to returning a copy of the tree
in BaseGraph.render_tree