Dashed Line Option Added a dashed line option to the graph. This can be either one of the specified types in styles.css (which can be expanded as necessary). OR It can be a list representing the dash/space length as required.
For example, to use the 0th pre-set dash (even short dashes):
xy_chart.dashed_line=0
OR
for a new dash array
xy_chart.dashed_line=[1,2,3,4,5,6]
This create a dash of length 1, a gap of length 2, a dash of length 3, a gap of length 4, ...

This was already possible with manually specifying CSS (see: http://pygal.org/news/, config.css.append('inline: path { stroke-dasharray: 15, 5, 10, 5; } .guide { stroke: none !important }')), but this might be more user friendly.

This is intended for a later expansion of individual styles per line.