Created a new view class
CompleteLogView is able to choose vertical and horizontal orientation, set either x or y or both axises to log or linear progressions, and choose a logarithmic progression besides base 10. It is also possible to change the orientation of the view while switching the the orientation of the progressions or keeping them attached to the axises they originally were attached to. It can also change the progressions at any time.