continued work on completeLogView
Got rid of x_or_y method and replaced it with x and y methods. because they are more backwards compatible. I also added the checks for base 10 to those methods so that base 10 values will be marginally more accurate. I cut out some unnecessary logic in the init method. and refined the logic in the xlog and ylog setter methods.