Official abbreviations
Would be a better choice to use official abbreviations for cantons names.
Really difficult to use "litteral names" in Switzerland. For example: "Waadt" is writter in German but "Vaud" is the french name of this french speaking canton. "Bern" is in German, but "Berne" is the french name of this *bilingual* canton.

Ref: Official swiss admin website: http://www.bfs.admin.ch/bfs/portal/fr/index/dienstleistungen/premiere_visite/03/03_02.html