Adds render_to_html() method
This allows to render a graph to an html string.

It does exactly the same as render_in_browser(), but it returns
the string instead of opening the result in the browser.

This is useful if you need to render the graph to html
to embed it in a webpage.